# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Classroom Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /subjects:
    get:
      operationId: getSubjects
      tags: [subjects]
      summary: Get all subjects for current staff
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
    post:
      operationId: createSubject
      tags: [subjects]
      summary: Create a new subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
      responses:
        "201":
          description: Subject created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
  /subjects/{id}:
    put:
      operationId: updateSubject
      tags: [subjects]
      summary: Update a subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubjectRequest"
      responses:
        "200":
          description: Subject updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
    delete:
      operationId: deleteSubject
      tags: [subjects]
      summary: Delete a subject
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subject deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /timetable/{classId}:
    get:
      operationId: getTimetable
      tags: [timetable]
      summary: Get timetable for a class
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timetable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timetable"
  /timetable/{classId}/generate:
    post:
      operationId: generateTimetable
      tags: [timetable]
      summary: Generate timetable for a class (HOD only)
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTimetableRequest"
      responses:
        "200":
          description: Timetable generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Timetable"
  /timetable/staff/me:
    get:
      operationId: getStaffTimetable
      tags: [timetable]
      summary: Get current staff's weekly timetable
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Staff timetable
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimetableSlot"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    Subject:
      type: object
      properties:
        id:
          type: string
        staffId:
          type: string
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [THEORY, LAB]
        credits:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - staffId
        - name
        - code
        - type
        - credits
    CreateSubjectRequest:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [THEORY, LAB]
        credits:
          type: integer
      required:
        - name
        - code
        - type
        - credits
...
    TimetableSlot:
      type: object
      properties:
        id:
          type: string
        day:
          type: string
          enum: [MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY]
        startTime:
          type: string
        endTime:
          type: string
        subjectId:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
        subjectType:
          type: string
          enum: [THEORY, LAB]
        room:
          type: string
        staffId:
          type: string
        staffName:
          type: string
        classId:
          type: string
        className:
          type: string
      required:
        - id
        - day
        - startTime
        - endTime
        - subjectName
        - subjectCode
        - room
        - staffName
        - classId
        - className
    Timetable:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        className:
          type: string
        version:
          type: integer
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimetableSlot"
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - classId
        - className
        - version
        - slots
    GenerateTimetableRequest:
      type: object
      properties:
        labStartTime:
          type: string
          description: Lab start time (HH:MM format)
        labEndTime:
          type: string
          description: Lab end time (HH:MM format)
        teaBreakStart:
          type: string
        teaBreakEnd:
          type: string
        lunchBreakStart:
          type: string
        lunchBreakEnd:
          type: string
        maxContinuousHours:
          type: integer
          default: 2
      required:
        - labStartTime
        - labEndTime
...