# lib/api-zod/src/generated/types/timetableSlot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Classroom Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { TimetableSlotDay } from "./timetableSlotDay";
import type { TimetableSlotSubjectType } from "./timetableSlotSubjectType";

export interface TimetableSlot {
  id: string;
  day: TimetableSlotDay;
  startTime: string;
  endTime: string;
  subjectId?: string;
  subjectName: string;
  subjectCode: string;
  subjectType?: TimetableSlotSubjectType;
  room: string;
  staffId?: string;
  staffName: string;
  classId: string;
  className: string;
}